+++ PricingData/src/main/scala/cs4624/prices/StockPrice.scala
@@ -5,5 +5,5 @@ import java.time.Instant
 /**
   * Created by joeywatts on 2/12/17.
   */
-case class StockPrice(symbol: String, time: Instant, price: Double)
+case class StockPrice(symbol: String, time: Instant, price: BigDecimal)
